Enable OCR  - Improve Extract for Byte Arrays
For Tesseract OCR to be called from Tika, the path has to be the application path or the path has to be provided. This change provide TesseractPath support in the TextExtractor class.

In addition the Mime type and File information for a byte array can be passed in to assist with the detector determining the correct parser to use, some files like RFC822 (email) do not identify well using Magic detection on the stream.